+++ Modules/Tournament/Providers/TournamentServiceProvider.php
@@ -3,6 +3,10 @@
 namespace Modules\Tournament\Providers;
 
 use Illuminate\Support\ServiceProvider;
+use Livewire\Livewire;
+use Modules\Tournament\Livewire\TournamentList;
+use Modules\Tournament\Livewire\TournamentCreate;
+use Modules\Tournament\Livewire\TournamentDetail;
 
 class TournamentServiceProvider extends ServiceProvider
 {
@@ -12,6 +16,14 @@ class TournamentServiceProvider extends ServiceProvider
     public function boot(): void
     {
         $this->registerViews();
+        $this->registerLivewireComponents();
+    }
+
+    protected function registerLivewireComponents(): void
+    {
+        Livewire::component('tournament.tournament-list', TournamentList::class);
+        Livewire::component('tournament.tournament-create', TournamentCreate::class);
+        Livewire::component('tournament.tournament-detail', TournamentDetail::class);
     }
 
     public function register(): void